Fix circular module imports bug
