+++ backend/server/data_processors/feature_calculation.py
@@ -5,7 +5,7 @@ import pandas as pd
 import numpy as np
 
 from server.types import DataFrameTransformer
-from server.ml_models.data_config import AVG_SEASON_LENGTH
+from server.data_config import AVG_SEASON_LENGTH
 
 DataFrameCalculator = Callable[[pd.DataFrame], pd.Series]
 Calculator = Callable[[Sequence[str]], DataFrameCalculator]
