+++ backend/server/data_processors/team_data_stacker.py
@@ -2,7 +2,7 @@ from typing import List, Callable
 import pandas as pd
 import numpy as np
 
-from server.ml_models.data_config import INDEX_COLS
+from server.data_config import INDEX_COLS
 
 REQUIRED_COLS: List[str] = ["home_team", "year", "round_number"]
